requirements.txt
dnspython==2.4.*
orjson==3.9.*
